refactor: Split off AclInitializer.
