+++ src/init/Initializer.ts
@@ -1,0 +1,3 @@
+import { AsyncHandler } from '../util/AsyncHandler';
+
+export abstract class Initializer extends AsyncHandler {}
